// src/schemas/productsSearchByEAN.ts
import { z } from "zod";

export const productsSearchByEAN = z.object({
  theme_name: z.string(),
  name: z.string(),
  unit: z.string(),
  unit_price: z.array(z.string()),
  price: z.string(),
  promotion_valid_period: z.null(),
  details: z.string(),
  limit_message: z.unknown(),
  only_with_biedronka_card: z.boolean(),
  regular_price: z.string(),
  is_promotion: z.boolean(),
  is_minimal_basket_value_bonus_buy: z.boolean(),
  discount: z.unknown(),
  price_tag_info: z.string(),
  omnibus_price: z.string(),
  discount_to_omnibus_price: z.null(),
  omnibus_type: z.unknown(),
  description: z.string(),
  type_of_wine: z.string(),
  sweetness_of_wine: z.string(),
  country_of_origin: z.string(),
  product_additional_country_of_origin_info: z.boolean(),
  fruits_or_vegetables_class: z.string(),
  image_url: z.string(),
  thumb_url: z.string(),
  multipack: z.unknown(),
  is_multipack: z.boolean(),
});